% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMCbtadjust-package.R
\docType{package}
\name{runMCMCbtadjust-package}
\alias{runMCMCbtadjust}
\alias{runMCMCbtadjust-package}
\title{The version 1.0.0 of this package was supported by the GAMBAS project funded by the French National Research Agency (ANR-18-CE02-0025).
This package was developed as part of the author activity at UR EFNO, INRAE, France.}
\description{
The function runMCMC_btadjust() returns a mcmc.list object which is the output of a Markov Chain Monte Carlo obtained - from either 'JAGS', 'nimble' or 'greta' - after adjusting burn-in and thinning parameters to meet pre-specified criteria in terms of convergence & effective sample size.
}
\author{
\strong{Maintainer}: Frédéric Gosselin \email{frederic.gosselin@inrae.fr} (\href{https://orcid.org/0000-0003-3737-106X}{ORCID}) (INRAE)

Other contributors:
\itemize{
  \item  Institut national de recherche pour l'agriculture, l'alimentation et l'environnement (INRAE) [copyright holder]
}

}
\keyword{internal}
